+++ noodles-tabix/src/io.rs
@@ -2,5 +2,6 @@
 
 pub mod indexed_reader;
 mod reader;
+mod writer;
 
-pub use self::reader::Reader;
+pub use self::{reader::Reader, writer::Writer};